Add old Generics tests as pretty test. Note that a few test cases are deactivated because we're emitting extra casts.
